Refactor attribute validation and performance testing scripts
This commit removes the old `validate_attribute.py` script and introduces a new `validate_attribute.py` for improved attribute search validation. Additionally, a new `performance_test_attribute.py` script is added to conduct performance tests on the attribute search strategy, evaluating speed, resource usage, and scalability under various load conditions. These changes enhance the testing framework and ensure more robust validation and performance metrics for the attribute search functionality.
